In the entire solution:   - Changed the type of board size local variables, method parameters and properties from byte to int   - Adjusted all their usage, among other things in converters, castings, error messages, other public constants/static, etc.

+++ NQueen.GUI/Views/MainView.xaml.cs
@@ -79,7 +79,7 @@ public partial class MainView : Window, IDisposable
 
             // Update the squares with the new dimensions
             viewModel.Chessboard.CreateSquares(
-                (byte)viewModel.BoardSize,
+                viewModel.BoardSize,
                 viewModel.Chessboard.Squares,
                 e.NewSize.Width,
                 e.NewSize.Height);
